feat: move parse logic from StreamScan Executor to source::ChunkReader
This PR moves the parse logic (e.g. JSON parsing) from the StreamScan Executor to the ChunkReader.

The ChunkReader is an intermediate layer between the Source and the Executor, reading the SourceMessage from the SourceReader by calling next() one by one, parsing it according to RowFormat, and generating a DataChunk for use by downstream methods.

It is important to note that the Source is a stream, and the user cannot know whether the stream has ended or not.
Currently, when using Kafka as Source, the configuration enable.partition.eof = true is used, so that the temporary absence of messages will be known by EOF error, this determines when the OLAP StreamScan can be finished. The logic for Streaming may need to be discussed to decide.