feat: Add epoch in StateStore::ingest_batch
* feat: Add epoch in StateStore::ingest_batch

This PR modifies the ingest_batch interface in StateStore to take the epoch.
The epoch is served as:
- A handle to represent an atomic write session. All ingested write batches
  associated with the same epoch have the all-or-nothing semantics,
  meaning that partial changes are not queryable and will be rollbacked if
  instructed.
- A version of a kv pair. kv pair associated with larger epoch is
  guaranteed to be newer then kv pair with smaller epoch.

Regardless of what the storage backend is, as long as we need the rollback
semantics described in 1), we will need the epoch. Therefore, we choose to
put the epoch in the StateStorage interface, not just in a specific backend
implementation.