test(unit): Add a test case for PubsubConfig
Since authentication is not being tested in any tests currently,
this is a minimal public contract test to see the config sets the
environment for google-cloud-pubsub as expected. Ideally we also
want tests for the From<SourceMessage> trait impl but that doesn't
appear possible at the moment due to inability to construct a mock
pubsub ReceivedMessage.