feat: Google Pubsub Connector (#5476)
* deps: Add pubsub crate

* feat: implement into<SourceMessage> for pubsub messages

* feat: implement split enumerator

* chore: Fix linting concerns

* fix: pubsub SourceMessage transformer

- pubsub message type: google_cloud_pubsub returns its own type
- ack_id is not yet exposed: linked tracking upstream issue

* test: remove pubsub message -> source message test

* feat: Implement pubsub split reader

* chore: Add comments for future iteration

* fix: Assign identifiers to splits

* chore: Remove redundant serde aliases

* chore: fmt

* chore: Cleanup pubsub properties

- Removes topic as configuration option as discussed.
  Our consumption primitive is a subscription, which
  we'll require to be made for us before we consume.
- Adds the `credentials` field.

* feat: ack received messages

* fix: Remove subscription creation precedure

As discussed, subscriptions must be created by the user and fed as a
source to RisingWave. We don't manipulate the subscription in any way.

* chore: set pubsub emulator host

required by the google-cloud-pubsbu crate to use an emulator

* chore: Remove attribute filter config.

Attribute-based filtering is configured on a subscription.
Since creating a subscription is not within our purview
anymore, configuring a filter falls back to the user.

* fix: Cleanup removed properties

* fix: Cleanup implementations and tag messages

- Cleanup implementation composition (remove unneded properties)
- Tag messages with origin splits and offsets (timestamps)

* feat: link pubsub connector to source crate

Also some fmt and refactor

* feat: dummy impl split.copy_with_offset

* fix: s/pub/pub(crate) (import)

* chore: Add copyright headers

* chore: linting concerns

* deps: pubsub with seek and auth from env

updates pubsub crate for env auth and seek support

* feat(pubsub): seek to offset

! Broken implementation owing to the fact that copy_with_offset for a
pubsub split needs to be async to call
google_cloud_pubsub::Subscription.seek but the split
implementation macro expects a non-async implementation. Will
cascade upwards.

* fix: Move seek op to PubsubReader

Discussions taught me any state-mutating effects should go to the
SplitReader while the Split itself should be a pure metadata reflection
on the state. With this commit the seek op is moved to the PubsubReader
impl.

* chore: Improve error handling

- Replaces unwraps/expects with standard result bubble.
- Gets rid of a utlity to_anyhow error tranformer in favor
  of an inline closure as with the rest of the codebase.
- Some cleanup

* refactor: pubsub reader to async-stream

In line with refactoring from #5647, this commit refactors
the pubsub reader impl to async-streams

* chore: update cargo lock

Auto-update cargo lock, address a linting concern and
add a comment.

* chore: cleanup

* fix: Pubsub properties (split_count)

As we get all attributes as strings from
postgres wire, the split count needed a
serde_as/with to unmarshall to an
numeric form from a string.

* chore: Improve pubsub split identifiers

* fix: duplicate pubsub prop

Renames a duplicate serde alias for the credentials
prop. Adds documentation to fields.

* fix: remove pubsub reader sleep

conditional sleep not needed with the pubsub connector as discussed

* fix: tag messages with valid splitids

* chore: address linting concerns

Removes a redundant .into() and fixes formatting.

* fix: Validate config in split enumerator

* docs: Enhance documentation of pubsub properties

* docs: fix pubsub rustdocs

Start rustdocs with /// and improve documentation
at places
Also some cleanuo

* feat(risedev): Add pubsub service

Adds a pubsub service to risedev in line with the Kafka service

* feat: validate subscription in emumerator

Catches invalid subscriptions/typos in the validation
phase.

* test(e2e): Add pubsub setup script

Also fmt and other chores, hopefully clean.

fix: bad dep constraint

* test(e2e): Add slt test script for pubsub

* chore: Address linter concerns

* ci: Add pubsub service to docker-compose

Also sneaks in slt cleanup

* ci: fix pubsub emulator setup condition

* ci: fix pubsub service and checks

* feat: Create pubsub source with offset or snapshot

- Support starting from a timestamp or pubsub snapshot. Snapshots
  are pub/sub's guaranteed-retention idiom.
- Verify `retain_acked_messages` policy on subscription (in question).

* feat: stop_offset for pubsub splits

* test(e2e): patch regression

Removes start_offset for now, but probably
should be covered at some point. Setting it up
might be challenging as it needs to be dynamically
generated.

* chore: Address concerns from review

- Indents in pubsub's slt
- Use bail! in lieu of return Err(...)

* test(unit): Add a test case for PubsubConfig

Since authentication is not being tested in any tests currently,
this is a minimal public contract test to see the config sets the
environment for google-cloud-pubsub as expected. Ideally we also
want tests for the From<SourceMessage> trait impl but that doesn't
appear possible at the moment due to inability to construct a mock
pubsub ReceivedMessage.

* test(e2e): Add sanity-check case for configuration

* chore: cleanup comment

As discussed with @tabversion, we will enforce the retain-on-acked
configuration to enable the seek operation.

* fix: Adress linter concerns

* refactor: Improve reader error handling

The reader now bails on authentication errors but
keeps going on irrelevant/temporal errors.

* refactor: Replace let-else with match

* chore: Update usage of deprecated chrono methods

* test(e2e): Fix runner

Regression caused by running CDC tests with connectors

Co-authored-by: Bohan Zhang <tabvision@bupt.icu>