fix(compaction): fix compaction not trigger when flush sst too large (#6745)
fix https://github.com/risingwavelabs/risingwave/issues/6896

The compaction may not trigger when the flush sstable is too large because we limit the write-amplification. If we limit both write-amplification and compaction size, we can not meet both of them so that the size of L0 become larger and larger but meta would not trigger a compaction task.


Approved-By: Li0k

Co-Authored-By: Little-Wallace <bupt2013211450@gmail.com>