feat(now): frontend part of now expression (#6699)
- Summarize your change (**mandatory**)
Implement happy path of frontend part of now expression in order that optimizer can generate `NowExecutor` node.
- How does this PR work? Need a brief introduction for the changed logic (optional)
We modify `to_stream` method of `LogicalFilter`. It will convert `NowExpr` in filter expression into `NowNode` and convert the `LogicalFIlter` into a `DynamicFilter`.
- Describe clearly one logical change and avoid lazy messages (optional)
`LogicalFilter` with `now()` in filter expression will be replaced with a `DynamicFilter` with a `NowNode`.
- Describe any limitations of the current code (optional)
We haven't limited `now()` expression format.


Approved-By: st1page

Co-Authored-By: Liang Zhao <liang@singularity-data.com>