fix(sink): sink exec get params from fe (#6922)
In the past, we assume the SinkExecutor's input must be a Materialize executor. SinkExecutor in Compute Node depends on that and uses its upstream input's schema and pk_indices to build the sinkImpl. But after we support creating a sink with query, we can not simply use the upstream executor's schema and pk_indices as the sink's property.
it is just a workaround fix for this release. It will be like https://github.com/risingwavelabs/risingwave/pull/6899 finally.


Approved-By: BugenZhao
Approved-By: yuhao-su

Co-Authored-By: st1page <1245835950@qq.com>