fix: quick fix for high barrier latency (#6943)
See #6571 for background.

After some discussion with @BugenZhao @yezizp2012 @tabVersion, we realized that it's futile to attempt to "backpressure" the streaming jobs by limiting `in_flight_barrier_nums`. On the contrary, it makes things more tricky, such as the interval between barriers becoming much larger, which in turn results in uncontrolled memory consumption.

This PR removes the limit of `in_flight_barrier_nums` in a quick & dirty way, and reduces the permits of exchange channels to mitigate the high-latency issue.

Of course, we must remove the related code. Just do this fix for the imminent version release.


Approved-By: BugenZhao
Approved-By: yezizp2012

Co-Authored-By: Eric Fu <eric@singularity-data.com>