feat: addition of container aware resource utility functions (#6762)
This PR introduces container aware resource utility functions to Risingwave. e.g total memory available to Risingwave should return the correct limit regardless of whether running in a container or a vm. To implement this, the Cgroup interface files, which are used for docker containerisations are read for the limits and usage. If the file mount does not exist or value cannot be read then the default system memory is returned. The cgroup file mount is when running in a container is assumed to be at "/sys/fs/cgroup" by default at the moment, which should work in most if not all cases.


Approved-By: wangrunji0408

Co-Authored-By: sixletters <e0406779@u.nus.edu>
Co-Authored-By: sixletters <62091272+sixletters@users.noreply.github.com>