feat(batch): add `ordered` property for seq scan (#6959)
This PR adds the `ordered` property for seq scan. If the optimizer does not require the order of seq scan, we can utilize the optimization in #6927.

Rough test: this PR reduces the latency of selecting an empty table from 30ms to 15ms under `dev` and 4.5ms to 2ms under `release`.


Approved-By: xxchan
Approved-By: chenzl25

Co-Authored-By: Bugen Zhao <i@bugenzhao.com>