fix(storage): fix panics caused by merge iter future cancellation (#6985)
We experienced many bugs due to cancelling the future of merge iterator `next`. The bug is caused from, when we cancel and drop the future holding a `PeekMut` of heap, dropping the `PeekMut` will cause heap comparison, and may attempt to call `key()` on some iterators with intermediately inconsistent states.

Previous we fix the bugs by trying to avoid these intermediately inconsistent states, but it seems hard to cover all the inconsistent states and say that we have fixed this issue totally.

In this PR, we fix this issue from its root cause. We introduce a guard wrapping over the `PeekMut`. When this guard is dropped, we call `PeekMut::pop` on the `PeekMut` node, which avoids the heap comparison. In this way, even if the future is dropped on some inconsistent states, we will not read the `key` from these inconsistent states.

Note: with this `PeekMutGuard`, we will have to manually drop the guard carefully in every branch of our code. We can now manually drop the guard via `pop` or `used`.


Approved-By: Little-Wallace
Approved-By: hzxa21

Co-Authored-By: William Wen <william123.wen@gmail.com>