refactor(storage): remove iter prev method from streaming state table (#7078)
Currently, the `iter_prev_epoch` method is only used delta join, but delta join is not used yet in our system, and is not maintained for a period. If we remove the `iter_prev_epoch` from state table, the interface of our state table will become clean, and there will not be MVCC requirement on the state table itself. It will only read the latest written date. After this refactor, we can further put the memtable down to the local state store, and make the read interface of our local state store a simple kv.

Approved-By: wcy-fdu
Approved-By: st1page
Approved-By: hzxa21

Co-Authored-By: William Wen <william123.wen@gmail.com>
Co-Authored-By: William Wen <44139337+wenym1@users.noreply.github.com>