refactor(state table): remove disable_sanity_check method from state table (#7181)
Our current state table provides method `disable_sanity_check` to disable the sanity check after the state table is initialized. However, all usages of this method is still under the initialization phase of streaming executor. As we are moving the memtable down to the local state store, the logic of sanity check will also be moved, and we will need to provide this unnecessary method in the trait of local state store. Therefore, in this PR, we remove the `disable_sanity_check` method from the state table, and provide extra methods to initialize state tables without sanity check.

Approved-By: wcy-fdu
Approved-By: hzxa21
Approved-By: soundOfDestiny