refactor(storage): store Bytes in memtable and return OwnedRow in state table iter (#7081)
In the future, we will move the memtable down to the local state store, which will return `Bytes` in its read interface. The current memtable stores `Vec<u8>`, and will involve memory allocation and memory copy when creating a static `Bytes`, which will introduce extra overhead. In this PR, we change to store `Bytes` in memtable so that we can create a static `Bytes` via a shallow copy.

Approved-By: hzxa21
Approved-By: Li0k
Approved-By: BugenZhao