feat: add `ANY/ALL/SOME(array)` operator to support `get_indexes` in sqlalchemy (#6717)
As the title.
These operators are used in the right side of binary operators which return boolean value. They require that there are some expressions in the parentheses, which return arrays. `SOME` is a synonym for `ANY`.
In PostgreSQL, there are some other operators named `ANY/ALL/SOME(subquery)`. They are not supported in this PR.

Approved-By: xiangjinwu

Co-Authored-By: Zehua Zou <41586196+HuaHuaY@users.noreply.github.com>
Co-Authored-By: Xiangjin <xiangjin@singularity-data.com>