feat(stream): introduce SortBuffer(Part I) (#7177)
To implement the [EMIT ON WINDOW CLOSE](risingwavelabs/rfcs/pull/30) semantics, we need to decouple the output and cleaning in SortExecutor. We're considering introducing a common component `SortBuffer` instead of a concrete `SortExecutor`. `SortBuffer` will manage a `StateTable` for buffering and export the deletion interface to its caller. It will be used in `WindowAgg` and `SessionWindow`, and may also include `TumbleWindow` and `HopWindow`.

Approved-By: st1page