fix(storage): fix reading unregistered table id (#7294)
Currently, when reading on a table id that has not been registered with `new_local` on global read, the code will panic because when we search for the potential read version to read, we run into `read_version_map.get(table_id).unwrap()`, and the `unwrap` cause the code to panic. In this PR, we fix by returning an empty vector when the table id is being read. We have added a new unit test to cover the case of empty read.

Approved-By: hzxa21