feat(frontend): cast/cmp with session timezone (#7243)
Support:
* Cast
* [X] date to timestamptz
* [X] timestamp to timestamptz
* [X] timestamptz to date
* [X] timestamptz to timestamp
* [X] timestamptz to time
* Comparison
* [X] date with timestamptz
* [X] timestamp with timestamptz
* Cast
* [X] timestamptz to string
* ~~We have implemented this to always use UTC because it is the basis of many things.~~
* Now we use the session timezone
* [X] string-**without-zone** to timestamptz (e.g. `'2022-10-01 12:00:00'::timestamptz`)

Missing:
* Arithmetic
* interval-**var** + timestamptz = timestamptz (e.g. `interval '1' day + '2022-03-13 09:00:00Z'::timestamptz`)
* timestamptz + interval-**var** = timestamptz
* timestamptz - interval-**var** = timestamptz

---

When doing so, we will emit the PgResponse NOTICE:
```
dev=> select '2022-01-01 00:00:00'::timestamp::timestamp with time zone;
NOTICE:  Your session timezone is UTC. It was used in the interpretation of timestamps and dates in your query. If this is unintended,
change your timezone to match that of your data's with `set timezone = [timezone]` or
rewrite your query with an explicit timezone conversion, e.g. with `AT TIME ZONE`.

timestamp
---------------------------
2022-01-01 00:00:00+00:00
(1 row)
```

Approved-By: fuyufjh
Approved-By: chenzl25
Approved-By: xiangjinwu