fix(storage): fix create index with incorrect properties (#7295)
When handle_create_index , we need inhert the table_properties from index_table instead of the context of `create_index` handle_args. Another problem, when we use point `get` function, we still need to deal the result key with min_epoch. Because `dist_key_hash` limits the unit test, so use prefix_hint uniformly instead of it.

**This section will be used as the commit message. Please do not leave this empty!**

Please explain **IN DETAIL** what the changes are in this PR and why they are needed:

- fix create index with incorrect properties
- fix function get with reteintion_seconds and unit-test
- reorg read_option with prefix_hint and refactor the logic of dist_key_hash
- avoid source state cleanup by reteintion_seconds

Approved-By: hzxa21
Approved-By: chenzl25