fix(frontend): Allow join key alignment on `IS NOT DISTINCT FROM`, check that eq key types are aligned in frontend (#7398)
Allow join key alignment on `IS NOT DISTINCT FROM`.

Previously join key is not aligned. This resulted in failure to create a hash join on unequal but implicitly castable types when using `IS NOT DISTINCT FROM`.

Catch error earlier when eq keys are not type aligned when converting join to batch/stream.

Approved-By: chenzl25