test(streaming): make epoch.inc works in tests (#7430)
Currently, many of our usage on `Epoch` in test uses `epoch.inc()`, for which we intend to advance the epoch to the next one. However, `epoch.inc()` will not touch the original `epoch`, but rather create a new epoch, which we didn't use in most tests, so in most streaming unit tests, the `epoch` is not changed, and many corner cases cannot be tested.

In this PR, we change to update the `epoch` when `epoch.inc()` is called. And the `commit_for_test` in state table, which bypass the epoch sanity check, is removed.

Approved-By: wcy-fdu