fix(stream agg): fix cache filling to avoid relying on `row_count` (#7411)
Previously we have an in-memory `row_count` variable maintained by `apply_batch`, and the value is restored from `row_count` column in result table. However, the problem is that, when filters are applied on some agg calls (especially different filters for different calls), the count maintained by `apply_batch` will be inconsistent with the `row_count` column. So we decided to remove the use of `row_count` in state caches.

Approved-By: st1page
Approved-By: BugenZhao
Approved-By: soundOfDestiny