chore: check breaking change in protobuf (#7609)
This PR adds the workflow of detecting breaking changes in protobuf to the CI workflow. Here is an example of error reporting:
- https://github.com/risingwavelabs/risingwave/pull/7609/commits/c1dd48e4f4291270ec515886d999dbf8b946b733

We've set the rule category to `WIRE_JSON` which is explained [here](https://docs.buf.build/breaking/rules). Actually `WIRE` is enough for our major usages, while `WIRE_JSON` ensures that the name of the fields is not modified, which brings less confusion to the codebase and also makes the JSON-based dashboard easier to maintain (e.g., avoid #7262 from happening).

We may not need to add this workflow to GitHub's branch protection rule until the next release.

Approved-By: TennyZhuang