feat(test): load balancing using DNS and IPVS in simulation (#7580)
This PR utilizes the new DNS and IPVS feature in madsim to support load balancing in deterministic test.

As an example, we added a DNS record `frontend -> 192.168.2.0` to the cluster, and a virtual service `192.168.2.0:4566 -> {192.168.2.1:4566, 192.168.2.2:4566, 192.168.2.3:4566}` to each node. We could then connect a frontend using the address `frontend:4566` which will be load balanced to 3 frontend servers.

This can also be used to simulate k8s services when we have multiple meta services after a few days.

- [x] The patch for madsim in Cargo.toml will be replaced by new versions once all tests pass.

Approved-By: BugenZhao