refactor(storage): remove storage core from local hummock (#7642)
Currently, in the `LocalHummockStorage`, most of useful fields are inside an inner `HummockStorageCore`, which are wrapped in `Arc` and was intended to be a struct shared in multiple tokio tasks. However, inside the `HummockStorageCore`, there is a `Arc<RwLock<ReadVersion>>`, and this is the real struct that is shared. Therefore, there is no need to wrap the `HummockStorageCore` within `Arc`, and furthermore we can move the fields in `HummockStorageCore` to `LocalHummockStorage` and remove the `HummockStorageCore`.

We also remove `Clone` from `LocalHummockStorage` in unit tests, and some unit test code are refactored accordingly.

Approved-By: Li0k