feat(cli): fallback to env if CLI arg is absent (#7658)
As title.

All envs will have the prefix of `RW`. Clap does not support prefixing env at the moment, so we have to specify the name manually :hot_face:. Serfig provides this functionality, but it still has the limitation of being unable to override with default value.

As for backward compatibility, although the env names have changed, the CLI args remain the same.

Approved-By: fuyufjh