feat(storage): space reclaim compaction  (#7597)
Introduce independent SpaceReclaim and TtlReclaim capabilities to the existing Compaction process. Introduce the concept of task_type to differentiate the capabilities of existing CompactTask and refactor the pick_compact_task process to simplify the code.Reorganize the picker to make the code structure more clear.


**This section will be used as the commit message. Please do not leave this empty!**

Please explain **IN DETAIL** what the changes are in this PR and why they are needed:

- Reorganize the picker code
- SpaceReclaim and TtlReclaim picker
- refactor the pick_compact_task process with task_type

Approved-By: Little-Wallace
Approved-By: zwang28
Approved-By: hzxa21