feat(expr): support pow function (#7789)
As per the title.

Right now, both x and y are casted into Float64 when pow(x, y).
By PG's standard, we also need to support Decimal but Decimal has no built-in pow and we may also change the underlying crate for Decimal. So we leave it aside for now.

Urgently requested by user.

#7725

Approved-By: st1page
Approved-By: TennyZhuang