feat(storage): use local thread to report metrics (#7620)
close https://github.com/risingwavelabs/risingwave/issues/7371

Every time we report a value to metrics, we need to query a hashmap with rw-lock, and then update an atomic value. When our system run one a machine with several CPU cores it would cause cpu-cache miss because we update the atomic value frequently.
I create a thread local map for every table. So that we could only update the thread local variables and flush it in time

Approved-By: Gun9niR
Approved-By: Li0k