chore(risedev): update the version and script of connector node component (#7583)
Update the version and auto-download script of connector node component. Since the connector node is still private, we release it via another [public repo](https://github.com/risingwavelabs/risingwave-connector-release/).

Currently, only CDC source connectors are released to the repo. Because the size of our sink connectors is quite large (including a 200MB dependency). It will be time-consuming to download such a huge file.

Approved-By: tabVersion
Approved-By: fuyufjh
Approved-By: liurenjie1024