feat(streaming): support table version for DML (#7918)
This PR makes the DML manager and related executors table-version-awared, as a preparation for the schema change. The frontend is untouched, so all versions in compute node will be `INITIAL_TABLE_VERSION_ID` now.

When the DML executor of the new version of the table is activated by receiving its first barrier, the new DML reader will be registered and it overwrites the old one. During this phase, all requests with the old version will be rejected and get an error message of "schema changed, try again", until the frontend receives the `Update` message of the new table catalog through the notification service.

Related:
- #7909

Approved-By: xx01cyx