feat(meta): support schema change in actor graph builder (#7878)
This PR extends the new actor graph builder introduced in #7659 for replacing table plan, and adds the procedure of preparing the `TableFragments`. Note that the RPC with the compute nodes is not implemented so this is not utilized by the frontend yet.

Building an actor graph for schema change can be similar and symmetric to the MV on MV.
- For MV on MV, we have existing upstream `Materialize` fragments that give the requirements to the distribution of current `Chain`. We'll generate new dispatchers for the upstream `Materialize` in the result.
- For replacing table plan for schema change, we (may) have existing downstream `Chain` fragments that give the requirements to the distribution of current `Materialize`. We'll generate updates for the merger for the downstream `Chain` in the result.

Related:
- #7908

Approved-By: chenzl25
Approved-By: yezizp2012