feat: implement stddev/var function  (#7952)
This PR implements a simple variance and standard deviation function

Because Welford's online algorithm is difficult to implement in a streaming system, we use a textbook algorithm, the naive algorithm, which is simple enough to reuse agg operators such as sum, count, sum0, case, etc.

<img width="973" alt="image" src="https://user-images.githubusercontent.com/1997790/219010430-df2ff178-2967-4d4f-849b-6a9213b90e73.png">

Welford's online algorithm requires iterative updating of avg and variance, so if we need to use them we need to implement agg separately rather than using the existing agg such as sum/count.

Approved-By: st1page

Co-Authored-By: Peng Chen <peng@singularity-data.com>
Co-Authored-By: Shanicky Chen <peng@risingwave-labs.com>