fix(test): fix occasional timeout of recovery test (#8017)
This PR tries to fix the occasional timeout of recovery test #7979.

The timeout was caused by a deadlock in the cascade drop when killing a node. It is a bug in madsim.
I have fixed it in https://github.com/madsim-rs/madsim/pull/125 so this PR is simply a version bump.

This PR also contains an improvement to the etcd simulator inspired by today's PoC meeting.
It will raise an `etcdserver: request is too large` error when the request is larger than 1.5MB. https://github.com/madsim-rs/madsim/pull/124

Approved-By: yezizp2012