feat(expr): cast from `jsonb` to bool and number (#8008)
Support casting from `jsonb` scalar to smallint / int / bigint / decimal / real / double precision. It is a runtime error if the jsonb type does not match.

Note that we limit the number as `f64` but PostgreSQL uses arbitrary precision `numeric`. The `f64` is enough as suggested by RFC 8259 for good interoperability. To support higher precision, we need to enable `arbitrary_precision` of crate `serde_json` and exam related code path.

There is also a subtle difference between decimal and f64 in PostgreSQL, when they round `2.5` to integer. But this issue is currently shadowed by #5576.

Approved-By: BugenZhao