fix: workaround decode error caused by large key (#8064)
This is a hotfix to workaround decode error caused by store key larger than 64KB. In the future, we may come up with a better way to prevent large key from occuring in the storage (e.g. limit the state table primary key size).

Note that this change breaks the format compatibility but in a very limited manner (only if there are keys with exact len=64KB written prior to this PR).

Approved-By: soundOfDestiny
Approved-By: wcy-fdu
Approved-By: wenym1
Approved-By: Li0k