refactor: no need to resolve hanging channel (#8050)
The hanging channel is a channel that's external for a table fragments subgraph. We introduced this for MV on MV long ago, but after https://github.com/risingwavelabs/risingwave/pull/4045 and https://github.com/risingwavelabs/risingwave/pull/6170, this becomes not that necessary: we can simply create a pair of channels if not found in the `ChannelPool`.

https://github.com/risingwavelabs/risingwave/blob/b1891423a68b98dff39d7220a891cdea8bd67951/src/stream/src/task/mod.rs#L123-L138

This is a preparation step for #7908.

Approved-By: yezizp2012