feat(storage): Introduce TtlReclaimSelector and refactor the trigger logic of Scheduler (#7937)
Its part of https://github.com/risingwavelabs/risingwave/issues/6918

Improve and introduce TtlReclaimSelector, and introduce TtlReclaimTrigger to periodically initiate compaction against ttl for LastLevel to ensure that data can be reclaimed in a timely manner.  As more triggers are introduced, try to refactor the Scheduler's Trigger logic to ensure the maintainability of the code.
- Stream to simplify the code of the Scheduler trigger
- Replace the last_index policy with key_range to ensure that the compaction runs correctly

Approved-By: zwang28
Approved-By: Little-Wallace

Co-Authored-By: Li0k <yuli@singularity-data.com>
Co-Authored-By: Runji Wang <wangrunji0408@163.com>