fix(common): `ListArray::from_protobuf` expects wrong cardinality of inner array (#8091)
The internal of `ListArray` stores its data flattened. For example, `values (array[1]), (array[]::int[]), (null), (array[2, 3]);` stores an inner `I32Array` with `[1, 2, 3]`, along with offset array `[0, 1, 1, 1, 3]` and null bitmap `TTFT`.

The cardinality of this inner array is not the length of outer array, but the last element of offset array.

Fixes #8082

Approved-By: kwannoel

Co-Authored-By: Xiangjin <xiangjin@singularity-data.com>
Co-Authored-By: xiangjinwu <17769960+xiangjinwu@users.noreply.github.com>