refactor(sqlparser): use ident instead of keyword to parse row format variants (#8154)
This PR has modified how a 'ROW FORMAT' variant is parsed by replacing the use of keywords with identifiers. As @xiangjinwu suggested.
This approach is more sustainable as the number of row format variants can increase with product iteration, and it is also better to use a limited number of keywords. Keeping fewer differences from PostgreSQL keywords is always a good idea.

Approved-By: waruto210
Approved-By: liurenjie1024
Approved-By: xiangjinwu