ci: add back auto-retry for status 255
It was removed in #8852. Add back because it's too frequent and thus manual retry isn't friendly. I've checked some failures on recent PRs and only found 255, so did not add -1.

It *might be possible* that 255 is not "agent terminated", but I've never met it.

Only added retry for PRs first, because it caused most troubles. For other pipelines, maybe it's worth checking errors and retrying manually.

> Exited with status 255 (after intercepting the agent’s termination signal, sent because the agent was stopped)