build: add resolver=2 to workspace
Resolver 2 means dev-dependency are not unified.

https://rust-lang.github.io/rfcs/2957-cargo-features2.html#feature-unification

This needs to be opted-in. And there's a warning when it's not specified in latest toolchain:

warning: some crates are on edition 2021 which defaults to `resolver = "2"`, but virtual workspaces default to `resolver = "1"`
note: to keep the current resolver, specify `workspace.resolver = "1"` in the workspace root's manifest
note: to use the edition 2021 resolver, specify `workspace.resolver = "2"` in the workspace root's manifest