fix: openssl is compiled twice, once not vendored in linux
It’s broken since the workspace-hack change (#12961).

Openssl-sys is compiled twice on Linux, because features are not unified.

The one for build-dependency doesn’t have openssl-sys feature. It’s depended by tls-native, which is depended by sql-x, which is build-dependency in workspace-hack. (Since we didn't configure hakari to unify target os-specific dependencies, hakari unified it only *partialy*, and this cause them unnecessarily compiled twice.)

Simply put, it’s now a build-dependency so
- Feature is not unified. This can be fixed easily and is done in this PR.
- It will be compiled twice in release mode. Unless #12362, but it’s broken. Not a large problem(? Well, I’m not sure But it has been so). Just slower…

This only happens on linux, but not mac, because native-tls works differently for different os.

Temporary workaround this by manually unify feature in workspace-config to unblock daily pipeline and stuff. Later, we might delete [build-dependencies] for docker/release. Or completely get rid of workspace-hack.

TLDR: After this change, on Linux openssl is compiled once in debug mode, and still twice in release mode. Both once on mac.